fix(vercel): extract tickets rate limiter so route.ts has no non-Route exports
Next.js 16 strictly validates API route module exports against a closed
list (GET/POST/etc + runtime/dynamic/etc config). An extra export —
even a test helper — makes the build fail with:

  Route "src/app/api/tickets/route.ts" does not match the required
  types of a Next.js Route.
    "__resetRateLimitsForTests" is not a valid Route export field.

This broke Vercel preview deployments for every open PR since #225.

Move the in-memory bucket Map, consumeRateLimit(), DAILY_LIMIT, and the
test-only __resetRateLimitsForTests() into a sibling rate-limit.ts.
route.ts imports what it needs; the test imports __resetRateLimitsForTests
directly from rate-limit. No behaviour change.
